revertme: hack to make StringValue compile
Currently compose does not support default values with actual/expect functions.
When we use vararg it creates a default emptyArray if we do not provide an argument which causes the compose compiler to throw an ISE.
We can revert this commit once https://youtrack.jetbrains.com/issue/KT-44499 is merged.

Signed-off-by: Aditya Wasan <adityawasan55@gmail.com>