Add tablet specific UI (#686)
Hi @msfjarvis! Please have a look at what I have been able to implement
so far. When trying to call the ListDetail version in the MainActivity I
realized that the MainActivity is based on the BaseActivity class, and I
wasn't exactly sure how to set the parameters for ListDetail. Let me
know your thoughts when you get a chance. Thanks!

---------

Co-authored-by: Harsh Shandilya <me@msfjarvis.dev>